fix: Fix for `set_operations` of binary dtype
